+++ src/libYARP_OS/src/DgramTwoWayStream.cpp
@@ -38,7 +38,7 @@ using namespace yarp::os;
 #define WRITE_SIZE (60000-CRC_SIZE)
 
 
-static bool checkCrc(char *buf, int length, int crcLength, int pct,
+static bool checkCrc(char *buf, ssize_t length, ssize_t crcLength, int pct,
                      int *store_altPct = NULL) {
     NetType::NetInt32 alt = 
         (NetType::NetInt32)NetType::getCrc(buf+crcLength,(length>crcLength)?(length-crcLength):0);
@@ -67,7 +67,7 @@ static bool checkCrc(char *buf, int length, int crcLength, int pct,
 }
 
 
-static void addCrc(char *buf, int length, int crcLength, int pct) {
+static void addCrc(char *buf, ssize_t length, ssize_t crcLength, int pct) {
     NetType::NetInt32 alt = 
         (NetType::NetInt32)NetType::getCrc(buf+crcLength,
                                            (length>crcLength)?(length-crcLength):0);
@@ -460,7 +460,7 @@ void DgramTwoWayStream::closeMain() {
     happy = false;
 }
 
-int DgramTwoWayStream::read(const Bytes& b) {
+ssize_t DgramTwoWayStream::read(const Bytes& b) {
     reader = true;
     bool done = false;
     
@@ -478,7 +478,7 @@ int DgramTwoWayStream::read(const Bytes& b) {
 
             //YARP_ASSERT(dgram!=NULL);
             //YARP_DEBUG(Logger::get(),"DGRAM Waiting for something!");
-            int result = -1;
+            ssize_t result = -1;
             if (mgram && restrictInterfaceIp.isValid()) { 
                 /*
                 printf("Consider remote mcast\n");
@@ -493,7 +493,7 @@ int DgramTwoWayStream::read(const Bytes& b) {
                 result =
                     dgram->recv(readBuffer.get(),readBuffer.length(),dummy);
                 YARP_DEBUG(Logger::get(),
-                           String("MCAST Got ") + NetType::toString(result) +
+                           String("MCAST Got ") + NetType::toString((int)result) +
                            " bytes");
                 
             } else if (dgram!=NULL) {
@@ -503,7 +503,7 @@ int DgramTwoWayStream::read(const Bytes& b) {
                 result =
                     dgram->recv(readBuffer.get(),readBuffer.length(),dummy);
                 YARP_DEBUG(Logger::get(),
-                           String("DGRAM Got ") + NetType::toString(result) +
+                           String("DGRAM Got ") + NetType::toString((int)result) +
                            " bytes");
             } else {
                 onMonitorInput();
@@ -611,8 +611,8 @@ void DgramTwoWayStream::write(const Bytes& b) {
     Bytes local = b;
     while (local.length()>0) {
         //YARP_DEBUG(Logger::get(),"DGRAM prep writing");
-        int rem = local.length();
-        int space = writeBuffer.length()-writeAvail;
+        ssize_t rem = local.length();
+        ssize_t space = writeBuffer.length()-writeAvail;
         bool shouldFlush = false;
         if (rem>=space) {
             rem = space;
@@ -641,22 +641,22 @@ void DgramTwoWayStream::flush() {
     pct++;
 
     while (writeAvail>0) {
-        int writeAt = 0;
+        ssize_t writeAt = 0;
         //YARP_ASSERT(dgram!=NULL);
-        int len = 0;
+        ssize_t len = 0;
 
         if (mgram!=NULL) {
             len = mgram->send(writeBuffer.get()+writeAt,writeAvail-writeAt);
             YARP_DEBUG(Logger::get(),
                        String("MCAST - wrote ") +
-                       NetType::toString(len) + " bytes"
+                       NetType::toString((int)len) + " bytes"
                        );
         } else if (dgram!=NULL) {
             len = dgram->send(writeBuffer.get()+writeAt,writeAvail-writeAt,
                               remoteHandle);
             YARP_DEBUG(Logger::get(),
                        String("DGRAM - wrote ") +
-                       NetType::toString(len) + " bytes to " +
+                       NetType::toString((int)len) + " bytes to " +
                        remoteAddress.toString()
                        );
         } else {
